[x86_64] Stricter address types and kernel map
* Change address types to be `repr(transparent)`
* Use `u64` for addresses
* Add kernel memory map for x86_64, as it also needs to be used by the
bootloader

+++ x86_64/src/memory/paging/mod.rs
@@ -1,2 +1,2 @@
-pub const FRAME_SIZE: usize = 0x1000;
-pub const PAGE_SIZE: usize = 0x1000;
+pub const FRAME_SIZE: u64 = 0x1000;
+pub const PAGE_SIZE: u64 = 0x1000;
